+++ src/adapter/graphql/src/scalars/base64.rs
@@ -18,7 +18,6 @@ async_graphql::scalar!(
 );
 
 impl Base64Usnp {
-    #[expect(dead_code)]
     pub fn into_inner(self) -> bytes::Bytes {
         self.0
     }