+++ src/e2e/app/cli/repo-tests/src/rest_api/test_gql_dataset_adapters.rs
@@ -144,7 +144,7 @@ pub async fn test_gql_dataset_adapter_versioned_file_direct_upload_download_mt(
                 mutation finishUpload(
                   $datasetId: DatasetID!,
                   $uploadToken: String!,
-                  $extraData: JSON,
+                  $extraData: ExtraData!,
                   $expectedHead: Multihash,
                 ) {
                   datasets {